refactor: 페이지 컴포넌트 로직 관련 훅으로 분리
