+++ frontend/src/hooks/team/useTeamsCondition.ts
@@ -1,6 +1,16 @@
 import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+import useSnackbar from 'hooks/utils/useSnackbar';
+
+import { MESSAGE, ROUTES_PATH } from 'constants/constants';
 
 const useTeamsCondition = () => {
+  const { showSnackbar } = useSnackbar();
+  const navigate = useNavigate();
+
+  const accessToken = localStorage.getItem('accessToken');
+
   const [teamsCondition, setTeamsCondition] = useState({
     open: true,
     close: false,
@@ -22,11 +32,21 @@ const useTeamsCondition = () => {
     setTeamsCondition({ open: false, close: false, my: true });
   };
 
+  const handleClickTeamAddButton = () => {
+    if (accessToken) {
+      navigate(ROUTES_PATH.INTERVIEW_TEAMS_ADD);
+      return;
+    }
+
+    showSnackbar({ message: MESSAGE.NEED_LOGIN_SERVICE });
+  };
+
   return {
     teamsCondition,
     handleClickOpenTeamsButton,
     handleClickCloseTeamsButton,
     handleClickMyTeamsButton,
+    handleClickTeamAddButton,
   };
 };
